requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
install==1.3.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Naked==0.1.31
pycryptodome==3.9.9
PyYAML==5.3.1
requests==2.24.0
shellescape==3.8.1
urllib3==1.25.11
Werkzeug==1.0.1
